update user data to database
